Move configuration classes into a new package
