+++ src/main/java/org/mybatis/dynamic/sql/configuration/GlobalConfiguration.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.mybatis.dynamic.sql;
+package org.mybatis.dynamic.sql.configuration;
 
 import java.io.IOException;
 import java.io.InputStream;
